+++ py2df/classes/__init__.py
@@ -1,2 +1,2 @@
 from classes.utilities import *
-from classes.abc import *
+from classes.abcs import *
