+++ py2df/enums/enum_util.py
@@ -5,6 +5,7 @@ class AutoNameEnum(Enum):
     """
     An enum whose auto values are the respective names of the constants.
     """
+
     def _generate_next_value_(name, start, count, last_values):
         return name
 
@@ -13,6 +14,7 @@ class AutoLowerNameEnum(Enum):
     """
     An enum whose auto values are the respective *lowercase* names of the constants.
     """
+
     def _generate_next_value_(name, start, count, last_values):
         return name.lower()
 
@@ -21,5 +23,6 @@ class AutoUpperNameEnum(Enum):
     """
     An enum whose auto values are the respective *uppercase* names of the constants.
     """
+
     def _generate_next_value_(name, start, count, last_values):
         return name.upper()
